Implemented admin controller for trainer and added blocking/deleting

+++ symfony/src/Trainer/Service/TrainerManager.php
@@ -4,13 +4,18 @@ declare(strict_types=1);
 
 namespace App\Trainer\Service;
 
+use App\Admin\Entity\Admin;
+use App\RefreshToken\Repository\RefreshTokenRepository;
 use App\Trainer\DTO\AdminUpdateTrainerRequest;
 use App\Trainer\DTO\CreateTrainerRequest;
 use App\Trainer\DTO\UpdateTrainerRequest;
 use App\Trainer\Entity\Trainer;
 use App\Trainer\Repository\TrainerRepository;
 use App\TrainingType\Repository\TrainingTypeRepository;
+use DateTimeImmutable;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+use Symfony\Component\HttpKernel\Exception\ConflictHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
@@ -22,6 +27,7 @@ final readonly class TrainerManager
         private TrainerRepository $trainerRepo,
         private TrainingTypeRepository $trainingTypeRepo,
         private UserPasswordHasherInterface $passwordHasher,
+        private RefreshTokenRepository $refreshTokenRepo,
         private EntityManagerInterface $entityManager,
     )
     {}
@@ -62,7 +68,7 @@ final readonly class TrainerManager
         return $trainer;
     }
 
-    public function updateByAdmin(Trainer $trainer, AdminUpdateTrainerRequest $requestDto): Trainer
+    public function updateByAdmin(AdminUpdateTrainerRequest $requestDto, Trainer $trainer): Trainer
     {
         if ($requestDto->firstName !== null) {
             $trainer->setFirstName($requestDto->firstName);
@@ -120,11 +126,55 @@ final readonly class TrainerManager
         return $trainer;
     }
 
-    public function softDelete(Trainer $trainer): void
+    public function softDelete(Trainer $trainer, ?Admin $admin = null): void
     {
-        $this->trainerRepo->remove($trainer);
+        if ($admin !== null && $admin->getId() === $trainer->getId()) {
+            throw new AccessDeniedHttpException('You cannot delete yourself');
+        }
+
+        if ($trainer->getDeletedAt()) {
+            throw new ConflictHttpException("Trainer already deleted");
+        }
+
+        $this->entityManager->wrapInTransaction(function () use ($trainer) {
+            $this->trainerRepo->remove($trainer);
+
+            $this->refreshTokenRepo->removeAllByUser($trainer);
+        });
+    }
+
+    public function restore(Trainer $trainer): Trainer
+    {
+        $trainer->setDeletedAt();
 
         $this->entityManager->flush();
+
+        return $trainer;
+    }
+
+    public function block(Admin $admin, Trainer $trainer): Trainer
+    {
+        if ($admin->getId() === $trainer->getId()) {
+            throw new AccessDeniedHttpException('You cannot block yourself');
+        }
+
+        if ($trainer->getBlockedAt()) {
+            throw new ConflictHttpException('Trainer already blocked');
+        }
+
+        $trainer->setBlockedAt(new DateTimeImmutable());
+
+        $this->entityManager->flush();
+
+        return $trainer;
+    }
+
+    public function unblock(Trainer $trainer): Trainer
+    {
+        $trainer->setBlockedAt(null);
+        $this->entityManager->flush();
+
+        return $trainer;
     }
 
     public function countPrice(Trainer $trainer, int $durationMinutes): float
